backend/requirements.txt
flask
openai-whisper
moviepy